+++ .env
@@ -1,9 +1,10 @@
 SASS_PATH=node_modules:src/assets/styles
 REACT_APP_SENTRY_DSN=
 REACT_APP_SENTRY_ENVIRONMENT=
-REACT_APP_API_URI=https://venepaikka-api.test.hel.ninja/graphql_v2/
 REACT_APP_SERVICE_MAP_URI=https://palvelukartta.hel.fi/unit/
-TUNNISTAMO_API_TOKEN_ENDPOINT=https://tunnistamo.test.kuva.hel.ninja/api-tokens/
-REACT_APP_OIDC_AUTHORITY=https://tunnistamo.test.kuva.hel.ninja/
-REACT_APP_OIDC_CLIENT_ID=https://api.hel.fi/auth/berths-admin-ui
-REACT_APP_OIDC_SCOPE="openid https://api.hel.fi/auth/berths https://api.hel.fi/auth/helsinkiprofile"
+REACT_APP_API_URI=https://venepaikka-federation.test.kuva.hel.ninja/
+REACT_APP_TUNNISTAMO_URI=https://tunnistamo.test.kuva.hel.ninja
+REACT_APP_TUNNISTAMO_API_TOKEN_ENDPOINT=api-tokens
+REACT_APP_TUNNISTAMO_LOGOUT_ENDPOINT=logout
+REACT_APP_TUNNISTAMO_CLIENT_ID=https://api.hel.fi/auth/berths-admin-ui
+REACT_APP_TUNNISTAMO_SCOPE="openid https://api.hel.fi/auth/berths https://api.hel.fi/auth/helsinkiprofile"
