--- src/domain/auth/__mocks__/getAuthenticatedUser.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/camelcase
-export default () => Promise.resolve({ access_token: 'foo.bar.baz' });
